From home : change JpaRepository type to long

+++ src/main/java/com/booking/repository/RepRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.booking.bean.Rep;
 
-public interface RepRepository extends JpaRepository<Rep, Integer>{
+public interface RepRepository extends JpaRepository<Rep, Long>{
 
 	Rep findAllByRepid(Long repid);
